refactor: merge schemas
